Implement Package Resolver
